fix(node): update memory test imports for hooks module
