+++ crates/node/tests/memory.rs
@@ -1,7 +1,7 @@
 //! Integration tests for the memory system using InMemoryStorage (no disk I/O).
 
+use crabtalk_node::hooks::Memory;
 use std::sync::Arc;
-use tools::Memory;
 use wcore::{MemoryConfig, testing::InMemoryStorage};
 
 fn test_memory() -> Memory<InMemoryStorage> {
@@ -187,7 +187,7 @@ fn migration_converts_legacy_files() {
 
 #[test]
 fn slugify_examples() {
-    use tools::memory::entry::slugify;
+    use crabtalk_node::hooks::memory::entry::slugify;
 
     assert_eq!(slugify("Luna's Vet Appointment!"), "luna-s-vet-appointment");
     assert_eq!(slugify("hello world"), "hello-world");
@@ -199,7 +199,7 @@ fn slugify_examples() {
 
 #[test]
 fn entry_parse_roundtrip() {
-    use tools::memory::entry::MemoryEntry;
+    use crabtalk_node::hooks::memory::entry::MemoryEntry;
 
     let entry = MemoryEntry {
         name: "test-entry".to_owned(),
@@ -217,7 +217,7 @@ fn entry_parse_roundtrip() {
 
 #[test]
 fn bm25_tokenize() {
-    use tools::memory::bm25::tokenize;
+    use crabtalk_node::hooks::memory::bm25::tokenize;
 
     let tokens = tokenize("Hello, World! This is a test.");
     assert!(tokens.contains(&"hello".to_owned()));
@@ -230,7 +230,7 @@ fn bm25_tokenize() {
 
 #[test]
 fn bm25_score_ranks() {
-    use tools::memory::bm25::score;
+    use crabtalk_node::hooks::memory::bm25::score;
 
     let docs = vec![
         (0, "the weather is sunny today"),
